Add flexibility in date formatting for OpenArchives collections
This includes a additional setting Open Archives > Collections > Date Formatting to allow a choice to use the existing date formatting or use the date as is from the Open Archive. The SOLR schema needed to have the date field as a String instead of a Date to allow for this flexibility.